fix warnings for r1.8

+++ mindspore/core/ops/grad/acosh_grad.cc
@@ -52,7 +52,7 @@ AbstractBasePtr AcoshGradInfer(const abstract::AnalysisEnginePtr &, const Primit
                                const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
   auto prim_name = primitive->name();
-  const size_t input_num = 2;
+  const int64_t input_num = 2;
   (void)CheckAndConvertUtils::CheckInputArgs(input_args, kEqual, input_num, prim_name);
   auto types = AcoshGradInferType(primitive, input_args);
   auto shapes = AcoshGradInferShape(primitive, input_args);
