+++ mindspore/core/ops/grad/asin_grad.cc
@@ -50,7 +50,7 @@ AbstractBasePtr AsinGradInfer(const abstract::AnalysisEnginePtr &, const Primiti
                               const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
   auto prim_name = primitive->name();
-  const size_t InputNum = 2;
+  const int64_t InputNum = 2;
   (void)CheckAndConvertUtils::CheckInputArgs(input_args, kEqual, InputNum, prim_name);
   MS_EXCEPTION_IF_NULL(input_args[kInputIndex0]);
   MS_EXCEPTION_IF_NULL(input_args[kInputIndex1]);
